execution/dashboard/requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.23.0,<1.0.0
ruamel.yaml>=0.18.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
pytest>=7.0.0
httpx>=0.24.0
pytest-cov>=4.1.0
hypothesis>=6.0.0
